+++ lib/cocoapods-binary-cache/diagnosis/integration.rb
@@ -3,7 +3,7 @@ require_relative "base"
 module PodPrebuild
   class IntegrationDiagnosis < BaseDiagnosis
     def run
-      should_be_integrated = if Pod::Podfile::DSL.prebuild_job? \
+      should_be_integrated = if PodPrebuild.config.prebuild_job? \
                              then @cache_validation.hit + @cache_validation.missed \
                              else @cache_validation.hit \
                              end
@@ -12,7 +12,7 @@ module PodPrebuild
         module_name = spec(name)&.module_name || name
         framework_path = \
           @standard_sandbox.pod_dir(name) + \
-          PodPrebuild::Config.instance.prebuilt_path(path: "#{module_name}.framework")
+          PodPrebuild.config.prebuilt_path(path: "#{module_name}.framework")
         framework_path.exist?
       end
       Pod::UI.puts "🚩 Unintegrated frameworks: #{unintegrated}".yellow unless unintegrated.empty?