+++ lib/cocoapods-binary-cache/pod-binary/helper/prebuild_sandbox.rb
@@ -4,7 +4,7 @@ module Pod
   class PrebuildSandbox < Sandbox
     # [String] standard_sandbox_path
     def self.from_standard_sanbox_path(path)
-      prebuild_sandbox_path = Pathname.new(path).realpath + ".." + PodPrebuild::Config.instance.prebuild_sandbox_path
+      prebuild_sandbox_path = Pathname.new(path).realpath + ".." + PodPrebuild.config.prebuild_sandbox_path
       self.new(prebuild_sandbox_path)
     end