+++ lib/cocoapods-binary-cache/pod-binary/integration/alter_specs.rb
@@ -15,7 +15,7 @@ module Pod
         # as to compitable with older version and be less wordy.
         framework_file_path = target.framework_name
         framework_file_path = target.name + "/" + framework_file_path if targets.count > 1
-        framework_file_path = PodPrebuild::Config.instance.prebuilt_path(path: framework_file_path)
+        framework_file_path = PodPrebuild.config.prebuilt_path(path: framework_file_path)
         add_vendered_framework(spec, target.platform.name.to_s, framework_file_path)
       end
 
@@ -63,7 +63,7 @@ module Pod
         attributes["resources"] ||= []
         attributes["resources"] = [attributes["resources"]] if attributes["resources"].is_a?(String)
         attributes["resources"] += resource_bundle_names.map do |name|
-          PodPrebuild::Config.instance.prebuilt_path(path: "#{name}.bundle")
+          PodPrebuild.config.prebuilt_path(path: "#{name}.bundle")
         end
       end