+++ spec/prebuild/prebuild_installer_spec.rb
@@ -41,7 +41,7 @@ describe "Pod::PrebuildInstaller" do
       allow(@installer).to receive(:targets_to_prebuild).and_return(targets_to_prebuild)
       allow(sandbox).to receive(:exsited_framework_target_names).and_return([])
       allow(sandbox).to receive(:generate_framework_path).and_return(tmp_dir + "/Generated")
-      allow(Pod::Podfile::DSL).to receive(:prebuild_code_gen).and_return(prebuild_code_gen)
+      allow(PodPrebuild.config).to receive(:prebuild_code_gen).and_return(prebuild_code_gen)
     end
 
     it "runs code generation before building" do